src/requirements.txt
paho-mqtt==1.5.1
influxdb==5.3.1